Introduced new fields in model structure
Introduced new non-mandatory fields in model structure: rxnNotes,
rxnReferences, confidenceScores and metCharge. All relevant RAVEN
functions were modified for this purpose. Now it is possible to carry
all information for these aforementioned 4 fields when exporting the
model to Excel format and importing it back to MATLAB. However, the
data from these new fields is lost, when exporting the model with
exportModel function, so it is not recommended to use the latter
function.