Fixed directories problem
Now RAVEN can be put in miscellaneous directories, which comprises the
folders with special characters. Supported special characters in Unix
are ‘space’ and ‘!’. If these characters are found in the full path,
backslash is added before them. For Windows, all the folder names are
put between the double quotation marks.