Reverted several recent changes
The last commit labeled as “Fixed directories problem” was too
aggresive, since it involved modifying all RAVEN functions, which
utilised “fileparts” function. That is, the addition of escape
characters broke functionality of apache-poi and xml_toolbox. So the
corresponding functions were reverted back to their earlier versions.

Thus, the addition of escape characters in the directories are only
relevant in the functions, which utilise the external command line,
i.e. getBlast, getWoLFScores and getKEGGModelForOrganism.