Add note and confidence score getModelFromHomology
Now with rxnNotes and confidenceScores are now part of the model
structure, When a new draft model is made through getModelFromHomology,
all reactions are given (i) a confidence score of 2 ('based on
sequence'), and (ii) a note stating 'Reaction included by
getModelFromHomology'. This makes it easier to track development of a
model.