New changes in Excel format regarding grRules
From now on, gene associations in Excel format will have exactly the
same style as in model.grRules. This prevents to lose gene association
data in the cases, when OR and AND relationships are used in the same
reaction. The function importExcelModel will still recognise an old
gene association format and will import gene associations
appropriately. Meanwhile, the function exportToExcelFormat will export
model only to the new format.