exportModel: added FBCv2 support
exportModel now can be used to export RAVEN Matlab model structure into SBML Level 1 FBCv2 format. The function was depthly revised and updated that xml file would be generated by OutputSBML from libSBML, unlike the previous versions of exportModel. The function was updated using libSBML 5.15, so it is strongly recommended to use this version. In SBML Level 3 there is not noticeable difference between 'general' and 'cobra' dialects, so the option to export model to 'cobra' dialect was removed. The new option allowing to export gene complexes (see the function for more info) was added.