fix: Substitute error display by warnings
The error display was substituted by two different warnings, one in the case that the function is called inside of another RAVEN function and another one for when this is called as a stand-alone function.

Comments were improved and further instructions for potentially wrong grRules modification are now provided in the warnings.