fix: get the reversibility based on both `lb` and `ub` fields
In this commit, the reversibility is obtained by querying both `lb` and `ub` fields. This is to avoid the obsolete `rev` field and weird cases (e.g. ub=0 and lb=-1000). Note that all reactions in a model should be defined in their forward direction, so that their equations could be properly formulated by this function.