fix: rearrange RAVEN functions
Several RAVEN functions appearing in `core` directory seem to be more relevant to `external` and `plotting` folders.

The `external` directory should contain the functions which are related to external software (excluding SBML, solvers), KEGG or MetaCyc. In addition, to simplify the search for RAVEN functionalities, the `external` directory also includes the alternatives for several existing functions. For instance, functions `makeFakeBlastStructure` and `getBlastFromExcel` are alternative approaches to `getBlast`/`getDiamond`. Similarly, `parseScores` is an alternative to `getWoLFScores`. There are even more functions, which fit more to `external` than in `core` directory (e.g. `getExpressionStructure`, `parseTaskList`), but these function do not use or mimick the included external programs in any way, so it was decided not to move then into `external` directory.