feat: checkFunctionUniqueness detectes multiple RAVEN versions
Added the new feature to `checkFunctionUniqueness`: it now checks for multiple RAVEN versions in MATLAB path and gives a fatal error if it is the case. That is, if at least one quarter of the existing RAVEN functions are found somewhere else in MATLAB path, it is assumed that another RAVEN version exists in MATLAB path.