feat: revised the printed messages during KEGG reconstruction
The model reconstruction from KEGG, particularly the one which involves the search against KO specific HMMs, is a time consuming process. It is therefore seems relevant to revise the printed messages during this pipeline and make them more detailed where necessary. Such details allow the user to faster familiarize with the whole pipeline and also identify the most time consuming stages. However, this comes at the cost of the code readability due to the higher number of FPRINTF lines. The following implementations were introduced to the KEGG reconstruction functions:
1. The messages are now printed in the different way than before. Previously, the progress of the most steps was reported only upon the full completion, for instance "Completed matching to HMMs". In this implementation, the user sees two messages regarding the start and the finish of the each stage.
2. Added the intermediate progress reporting messages for: search against HMMs; HMMs training; clustering and multiple alignment; multi-FASTA files generation. The messages reporting about 25%, 50% and 75% completion are printed to the screen.
3. Added the reports for the stages, which were not reported through the messages before, although their runtimes are noticeably long. As an example, the status regarding the GPR rules construction is now printed out.
4. Revised the usage of "NOTE: ". Previously, it was used once importing kegg***.mat files. Now it is printed only during the non-critical cases, which may require the user's attention.
5. Added some cosmetic changes to the messages. This includes the title message when the model reconstruction begins and ends.
6. Other small changes, making the comments in the code more formal.